python/requirements.txt
aiofiles==22.1.0 ; python_version >= "3.9" and python_version < "4.0"
ansible-core==2.14.4 ; python_version >= "3.9" and python_version < "4.0"
ansible-runner==2.3.2 ; python_version >= "3.9" and python_version < "4.0"
ansible==7.4.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.9" and python_version < "4.0"
beautifulsoup4==4.12.2 ; python_version >= "3.9" and python_version < "4.0"
brotli==1.0.9 ; platform_python_implementation == "CPython" and python_version >= "3.9" and python_version < "4.0"
brotlicffi==1.0.9.2 ; platform_python_implementation != "CPython" and python_version >= "3.9" and python_version < "4.0"
cachetools==5.3.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.9" and python_version < "4"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
colorlog==6.7.0 ; python_version >= "3.9" and python_version < "4.0"
cryptography==40.0.2 ; python_version >= "3.9" and python_version < "4.0"
cssselect2==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
docutils==0.19 ; python_version >= "3.9" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.9" and python_version < "4.0"
fastapi==0.87.0 ; python_version >= "3.9" and python_version < "4.0"
fonttools[woff]==4.39.3 ; python_version >= "3.9" and python_version < "4.0"
google-api-core==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-core[grpc]==2.11.0 ; python_version >= "3.9" and python_version < "4.0"
google-api-python-client==2.85.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth-httplib2==0.1.0 ; python_version >= "3.9" and python_version < "4.0"
google-auth==2.17.3 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-secret-manager==2.16.1 ; python_version >= "3.9" and python_version < "4.0"
google-cloud-storage==2.8.0 ; python_version >= "3.9" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.9" and python_version < "4.0"
google-resumable-media==2.4.1 ; python_version >= "3.9" and python_version < "4.0"
google==3.0.0 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.59.0 ; python_version < "4.0" and python_version >= "3.9"
googleapis-common-protos[grpc]==1.59.0 ; python_version >= "3.9" and python_version < "4.0"
grpc-google-iam-v1==0.12.6 ; python_version >= "3.9" and python_version < "4.0"
grpcio-status==1.50.0 ; python_version < "4.0" and python_version >= "3.9"
grpcio==1.50.0 ; python_version < "4.0" and python_version >= "3.9"
gunicorn==20.1.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
html5lib==1.1 ; python_version >= "3.9" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
lockfile==0.12.2 ; python_version >= "3.9" and python_version < "4.0"
markdown2==2.4.8 ; python_version >= "3.9" and python_version < "4"
markupsafe==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
motor==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.1 ; python_version >= "3.9" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.9" and python_version < "4.0"
pexpect==4.8.0 ; python_version >= "3.9" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.9" and python_version < "4.0"
proto-plus==1.22.2 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.22.3 ; python_version < "4.0" and python_version >= "3.9"
ptyprocess==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.7 ; python_version >= "3.9" and python_version < "4.0"
pydyf==0.6.0 ; python_version >= "3.9" and python_version < "4.0"
pymongo==4.3.3 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.9" and python_version < "4.0"
pyphen==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
python-daemon==3.0.1 ; python_version >= "3.9" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
python-slugify==6.1.2 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
libs/reddevil-3.0.4-py3-none-any.whl ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.9" and python_version < "4"
resolvelib==0.8.1 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
setuptools==67.6.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version < "4.0" and python_version >= "3.9"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
soupsieve==2.4.1 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.21.0 ; python_version >= "3.9" and python_version < "4.0"
text-unidecode==1.3 ; python_version >= "3.9" and python_version < "4.0"
tinycss2==1.2.1 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.9" and python_version < "4"
uvicorn==0.19.0 ; python_version >= "3.9" and python_version < "4.0"
weasyprint==57.2 ; python_version >= "3.9" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
zopfli==0.2.2 ; python_version >= "3.9" and python_version < "4.0"
